Imported using usd-importer.

Publish parent: 434aa32759a9333c98196392221e447b0245fa2d
Changelog parent: 321973918812ae277037908bfc0e2360860db5ea

+++ sysdeps/mach/hurd/fchflags.c
@@ -25,7 +25,7 @@
 
 /* XXX should be __fchflags? */
 int
-fchflags (int fd, unsigned long int flags)
+fchflags (int fd, int flags)
 {
   error_t err;
